requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
black==23.3.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
Django==4.2.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
flatbuffers==23.5.9
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
jax==0.4.10
joblib==1.2.0
keras==2.12.0
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
ml-dtypes==0.1.0
mypy-extensions==1.0.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
termcolor==2.3.0
threadpoolctl==3.1.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.3.4
wrapt==1.14.1
